+++ meta_package_manager.7h.py
@@ -179,53 +179,60 @@ class HomebrewCask(Homebrew):
         Sample of brew cask output:
 
             $ brew cask list --versions
-            aerial 1.2beta5, 1.1
+            aerial 1.2beta5
             android-file-transfer latest
-            audacity 2.1.2-1453294898, 2.1.2
-            bitbar 1.9.1
-            chromium latest
-            firefox 47.0, 46.0.1, 46.0
-            flux 37.3, 37.2, 37.1, 36.8, 36.6
-            gimp 2.8.16-x86_64
-            java 1.8.0_92-b14
-            prey
-            ubersicht
+            audacity 2.1.2
+            bitbar 1.9.2
+            firefox 49.0.1
+            flux 37.7
+            gimp 2.8.18-x86_64
+            java 1.8.0_112-b16
 
             $ brew cask info aerial
             aerial: 1.2beta5
-            Aerial Screensaver
             https://github.com/JohnCoates/Aerial
-            /usr/local/Caskroom/aerial/1.2beta5 (0B)
-            https://github.com/caskroom/homebrew-cask/blob/master/Casks/aerial.rb
-            ==> Contents
-              Aerial.saver (screen_saver)
+            /usr/local/Caskroom/aerial/1.2beta5 (18 files, 6.6M)
+            From: https://github.com/caskroom/homebrew-cask/blob/master/Casks/aerial.rb
+            ==> Name
+            Aerial Screensaver
+            ==> Artifacts
+            Aerial.saver (screen_saver)
 
             $ brew cask info firefox
-            firefox: 47.0.1
+            firefox: 50.0.1
+            https://www.mozilla.org/firefox/
+            /usr/local/Caskroom/firefox/49.0.1 (107 files, 185.3M)
+            From: https://github.com/caskroom/homebrew-cask/blob/master/Casks/firefox.rb
+            ==> Name
             Mozilla Firefox
-            https://www.mozilla.org/en-US/firefox/
-            Not installed
-            https://github.com/caskroom/homebrew-cask/blob/master/Casks/firefox.rb
-            ==> Contents
-              Firefox.app (app)
+            ==> Artifacts
+            Firefox.app (app)
 
             $ brew cask info prey
-            prey: 1.5.1
-            Prey
-            https://preyproject.com
+            prey: 1.6.3
+            https://preyproject.com/
             Not installed
-            https://github.com/caskroom/homebrew-cask/blob/master/Casks/prey.rb
-            ==> Contents
-              prey-mac-1.5.1-x86.pkg (pkg)
+            From: https://github.com/caskroom/homebrew-cask/blob/master/Casks/prey.rb
+            ==> Name
+            Prey
+            ==> Artifacts
+            prey-mac-1.6.3-x86.pkg (pkg)
+            ==> Caveats
+            Prey requires your API key, found in the bottom-left corner of
+            the Prey web account Settings page, to complete installation.
+            The API key may be set as an environment variable as follows:
+
+              API_KEY="abcdef123456" brew cask install prey
 
             $ brew cask info ubersicht
-            ubersicht: 1.0.42
-            Übersicht
-            http://tracesof.net/uebersicht
+            ubersicht: 1.0.44
+            http://tracesof.net/uebersicht/
             Not installed
-            https://github.com/caskroom/homebrew-cask/blob/master/Casks/ubersicht.rb
-            ==> Contents
-              Übersicht.app (app)
+            From: https://github.com/caskroom/homebrew-cask/blob/master/Casks/ubersicht.rb
+            ==> Name
+            Übersicht
+            ==> Artifacts
+            Übersicht.app (app)
         """
         # `brew cask update` is just an alias to `brew update`. Perform the
         # action anyway to make it future proof.
@@ -259,11 +266,11 @@ class HomebrewCask(Homebrew):
             # Inspect the package closer to evaluate its state.
             output = self.run(self.cli, 'cask', 'info', name)
 
-            # Consider package as up-to-date if installed.
-            if output.find('Not installed') == -1:
-                continue
-
             latest_version = output.split('\n')[0].split(' ')[1]
+
+            # Skip already installed packages.
+            if version == latest_version:
+                continue
 
             self.updates.append({
                 'name': name,