+++ CHANGES.rst
@@ -8,6 +8,7 @@ Changelog
 * Sort list of outdated packages by lower-cased package IDs.
 * Fetch currently installed version from ``cask``.
 * Bump minimal requirement of ``brew`` and ``cask`` to 1.0.0.
+* Fix fetching of outdated ``cask`` packages.
 
 
 `1.10.0 (2016-10-04) <https://github.com/kdeldycke/meta-package-manager/compare/v1.9.0...develop>`_